@@ -1,6 +1,6 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFrameworks>netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
+		<TargetFrameworks>net6.0</TargetFrameworks>
 		<DefineConstants>NETCORE;</DefineConstants>
 		<DefineConstants>NETCOREAPP1_1</DefineConstants>
 		<PackageId>GeneXus.SDAPI.Core</PackageId>
