@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFrameworks>netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
+		<TargetFrameworks>net6.0</TargetFrameworks>
 		<DefineConstants>NETCORE</DefineConstants>
 		<PackageTags>DynService OData</PackageTags>
 		<PackageId>GeneXus.DynService.OData.Core</PackageId>
