@@ -11,7 +11,7 @@ namespace GeneXus.Mime
 		public static string GetMimeMapping(string fileName)
 		{
 			if (fileName == null)
-				throw new ArgumentNullException("fileName");
+				throw new ArgumentNullException(nameof(fileName));
 			return MimeMapping._mappingDictionary.GetMimeMapping(fileName);
 		}
 
@@ -335,6 +335,7 @@ namespace GeneXus.Mime
 				this.AddMapping(".stl", "application/vnd.ms-pki.stl");
 				this.AddMapping(".sv4cpio", "application/x-sv4cpio");
 				this.AddMapping(".sv4crc", "application/x-sv4crc");
+				this.AddMapping(".svg", "image/svg+xml");
 				this.AddMapping(".swf", "application/x-shockwave-flash");
 				this.AddMapping(".t", "application/x-troff");
 				this.AddMapping(".tar", "application/x-tar");
