@@ -61,6 +61,10 @@ namespace GeneXus.Utils
 				return null;
 			}
 		}
+		internal static bool IsNull(object instance)
+		{
+			return (instance == null || instance == NTSGeographyWrapper.NullSQLGeography);
+		}
 
 		internal static object STGeometryType(object instance)
 		{
@@ -649,22 +653,27 @@ namespace GeneXus.Utils
 			}
 			try
 			{
-				// Parse
 				_innerValue = NTSGeographyWrapper.Parse(geoText);
 				// SRID, Type & Points X, Y
-				if ((!NTSGeographyWrapper.IsValid(_innerValue)) && _innerValue != null)
+				if (_innerValue != null && !NTSGeographyWrapper.IsNull(_innerValue))
 				{
-					//_innerValue = NTSGeographyWrapper.MakeValid(_innerValue);
-				}
+					if (NTSGeographyWrapper.IsValid(_innerValue))
+					{
+						//_innerValue = NTSGeographyWrapper.MakeValid(_innerValue);	
+						this.srid = NTSGeographyWrapper.Srid(_innerValue);
+						this.setGXGeoType(NTSGeographyWrapper.STGeometryType(_innerValue).ToString());
 
-				this.srid = NTSGeographyWrapper.Srid(_innerValue);
-
-				this.setGXGeoType(NTSGeographyWrapper.STGeometryType(_innerValue).ToString());
-				if (GeographicType == GeoGraphicTypeValue.Point)
-				{
-					this.Point.Longitude = NTSGeographyWrapper.Long(_innerValue);
-					this.Point.Latitude = NTSGeographyWrapper.Lat(_innerValue);
+						if (GeographicType == GeoGraphicTypeValue.Point)
+						{
+							this.Point.Longitude = NTSGeographyWrapper.Long(_innerValue);
+							this.Point.Latitude = NTSGeographyWrapper.Lat(_innerValue);
+						}
+					}
+					else
+						setNullGeography();
 				}
+				else
+					setNullGeography();
 			}
 			catch (Exception ex)
 			{
@@ -676,22 +685,24 @@ namespace GeneXus.Utils
 					}
 					else
 					{
-						// Cannot parse value
-						_innerValue = NTSGeographyWrapper.NullSQLGeography;
-						this.geoText = "";
-						this.Point.Longitude = 0;
-						this.Point.Latitude = 0;
+						setNullGeography();
 					}
 				}
 				else
 				{
-					// Cannot parse value
-					_innerValue = NTSGeographyWrapper.NullSQLGeography;
-					this.geoText = "";
-					this.Point.Longitude = 0;
-					this.Point.Latitude = 0;
+					setNullGeography();
+					
 				}
 			}
+		}
+
+		void setNullGeography()
+		{
+			// Cannot parse value
+			_innerValue = NTSGeographyWrapper.NullSQLGeography;
+			this.geoText = "";
+			this.Point.Longitude = 0;
+			this.Point.Latitude = 0;
 		}
 
 		override public String ToString()
@@ -930,7 +941,6 @@ namespace GeneXus.Utils
 			}
 			return ArrayOfPoints;
 		}
-
 	}
 
 	public class PointT
@@ -948,6 +958,5 @@ namespace GeneXus.Utils
 	{
 		public Line[] Points;
 	}
-
 
 }
