@@ -19,6 +19,8 @@ using TZ4Net;
 using GxClasses.Helpers;
 using System.Net;
 #endif
+using NUglify;
+using NUglify.Html;
 using GeneXus.Web.Security;
 
 using System.Web;
@@ -39,6 +41,8 @@ using System.Collections.Concurrent;
 using System.Drawing.Drawing2D;
 using GeneXus.Storage;
 using GeneXus.Services;
+using GeneXus.Http;
+using System.Security;
 
 namespace GeneXus.Utils
 {
@@ -59,10 +63,10 @@ namespace GeneXus.Utils
 		{
 			lock (random)
 			{
-				var bytes = new Byte[8];
+				byte[] bytes = new Byte[8];
 				random.GetBytes(bytes);
 
-				var ul = BitConverter.ToUInt64(bytes, 0) / (1 << 11);
+				ulong ul = BitConverter.ToUInt64(bytes, 0) / (1 << 11);
 				double d = ul / (double)(1UL << 53);
 				return d;
 			}
@@ -2632,19 +2636,19 @@ namespace GeneXus.Utils
 		public string Time()
 		{
 			return DateTime.Now.ToString(TimeFormatFromSize(8, -1, ":"), cultureInfo);
-        }
+		}
 
-        static public DateTime AddMth(DateTime dt, int cantMonths)
-        {
-            if (dt == nullDate)
-                return nullDate;
-            return dt.AddMonths(cantMonths);
-        }
-        static public DateTime AddYr(DateTime dt, int cantYears)
-        {
-            if (dt == nullDate)
-                return nullDate;
-            return dt.AddYears(cantYears);
+		static public DateTime AddMth(DateTime dt, int cantMonths)
+		{
+			if (dt == nullDate && (cantMonths < 0 || Preferences.IgnoreAddOnEmptyDates))
+				return nullDate;
+			return dt.AddMonths(cantMonths);
+		}
+		static public DateTime AddYr(DateTime dt, int cantYears)
+		{
+			if (dt == nullDate && (cantYears < 0 || Preferences.IgnoreAddOnEmptyDates))
+				return nullDate;
+			return dt.AddYears(cantYears);
         }
         static public DateTime DateEndOfMonth(DateTime dt)
         {
@@ -2693,30 +2697,29 @@ namespace GeneXus.Utils
 		}
         static public int DDiff(DateTime dtMinu, DateTime dtSust)
         {
-            return Convert.ToInt32((dtMinu - dtSust).TotalDays);
-        }
-        static public DateTime TAdd(DateTime dt, int seconds)
-        {
-            if (dt == nullDate)
-                return nullDate;            
-            return dt.AddSeconds(seconds);
-        }
-        static public DateTime TAddMs(DateTime dt, double seconds)
-        {
-
-            if (dt == nullDate)
-                return nullDate;
-            if (seconds % 1 == 0)
-                return dt.AddSeconds((int)seconds);
-            else
-                return dt.AddMilliseconds(seconds * 1000);
-        }
-        static public DateTime DAdd(DateTime dt, int days)
-        {
-            if (dt == nullDate)
-                return nullDate;
-            return dt.AddDays(days);
-        }
+			return Convert.ToInt32((dtMinu - dtSust).TotalDays);
+		}
+		static public DateTime TAdd(DateTime dt, int seconds)
+		{
+			if (dt == nullDate && (seconds < 0 || Preferences.IgnoreAddOnEmptyDates))
+				return nullDate;
+			return dt.AddSeconds(seconds);
+		}
+		static public DateTime TAddMs(DateTime dt, double seconds)
+		{
+			if (dt == nullDate && (seconds < 0 || Preferences.IgnoreAddOnEmptyDates))
+				return nullDate;
+			if (seconds % 1 == 0)
+				return dt.AddSeconds((int)seconds);
+			else
+				return dt.AddMilliseconds(seconds * 1000);
+		}
+		static public DateTime DAdd(DateTime dt, int days)
+		{
+			if (dt == nullDate && (days < 0 || Preferences.IgnoreAddOnEmptyDates))
+				return nullDate;
+			return dt.AddDays(days);
+		}
         public DateTime CToT(string strDate, int picFmt, int ampmFmt)
         {
             if (isNullDateTime(strDate, picFmt, ampmFmt))
@@ -2866,10 +2869,13 @@ namespace GeneXus.Utils
 			//DateTime dtRet = new DateTime(dt.Year,dt.Month,dt.Day,0,0,0,0);
 			//return dtRet; 
 			//This is more efficient than creating a new datetime..
-			dt = dt.AddMilliseconds(dt.Millisecond * -1);
-			dt = dt.AddSeconds(dt.Second * -1);
-			dt = dt.AddMinutes(dt.Minute * -1);
-			dt = dt.AddHours(dt.Hour * -1);
+			if (dt.Millisecond != 0 || dt.Second != 0 || dt.Minute != 0 || dt.Hour != 0)
+			{
+				dt = dt.AddMilliseconds(dt.Millisecond * -1);
+				dt = dt.AddSeconds(dt.Second * -1);
+				dt = dt.AddMinutes(dt.Minute * -1);
+				dt = dt.AddHours(dt.Hour * -1);
+			}
 			return dt;
 		}
 		public static DateTime ResetDate(DateTime dt)
@@ -2877,9 +2883,12 @@ namespace GeneXus.Utils
 			//DateTime dtRet = new DateTime(1,1,1,dt.Hour,dt.Minute,dt.Second);
 			//return dtRet;
 			//This is more efficient than creating a new datetime.
-			dt = dt.AddDays((dt.Day * -1) + 1);
-			dt = dt.AddMonths((dt.Month * -1) + 1);
-			dt = dt.AddYears((dt.Year * -1) + 1);
+			if (dt.Day != 1 || dt.Month != 1 || dt.Year != 1)
+			{
+				dt = dt.AddDays((dt.Day * -1) + 1);
+				dt = dt.AddMonths((dt.Month * -1) + 1);
+				dt = dt.AddYears((dt.Year * -1) + 1);
+			}
 			return dt;
 		}
 		public static DateTime ResetMilliseconds(DateTime dt)
@@ -2887,7 +2896,10 @@ namespace GeneXus.Utils
 			//DateTime dtRet = new DateTime(dt.Year,dt.Month,dt.Day,dt.Hour,dt.Minute,dt.Second,0);
 			//reutrn dtRet;
 			//This is more efficient than creating a new datetime.
-			dt = dt.AddMilliseconds(dt.Millisecond * -1);
+			if (dt.Millisecond != 0)
+			{
+				dt = dt.AddMilliseconds(dt.Millisecond * -1);
+			}
 			return dt;
 		}
 		public static DateTime ResetMillisecondsTicks(DateTime dt)
@@ -3520,10 +3532,17 @@ namespace GeneXus.Utils
 	}
 	public class PathUtil
 	{
+		const string schemeRegEx = @"^([a-z][a-z0-9+\-.]*):";
+		static Regex scheme = new Regex(schemeRegEx, RegexOptions.IgnoreCase);
+
 		public static bool IsAbsoluteUrl(string url)
 		{
 			Uri result;
 			return Uri.TryCreate(url, UriKind.Absolute, out result) && (result.Scheme == GXUri.UriSchemeHttp || result.Scheme == GXUri.UriSchemeHttps || result.Scheme == GXUri.UriSchemeFtp);
+		}
+		public static bool IsAbsoluteUrlOrAnyScheme(string url)
+		{
+			return (!String.IsNullOrEmpty(url)) && (IsAbsoluteUrl(url) || scheme.IsMatch(url));
 		}
 
 		public static bool HasUrlQueryString(string url)
@@ -3963,7 +3982,7 @@ namespace GeneXus.Utils
 			string value = parm.ToString();
 			if (!String.IsNullOrEmpty(gxkey))
 			{
-				string strValue = Crypto.Decrypt64(value.ToString(), gxkey);
+				string strValue = Crypto.Decrypt64(value.ToString(), gxkey, true);
 				if ((String.CompareOrdinal(StringUtil.Right(strValue, 6).TrimEnd(' '),
 					Crypto.CheckSum(StringUtil.Left(strValue, (short)(StringUtil.Len(strValue) - 6)), 6).TrimEnd(' ')) == 0))
 				{
@@ -4019,7 +4038,20 @@ namespace GeneXus.Utils
 			string number = padding + character.ToString();
 			buffer.Append("&#" + number + ";");
 		}
-
+		[SecuritySafeCritical]
+		public static string HTMLClean(string text)
+		{
+			HtmlSettings htmlSettings = new HtmlSettings { PrettyPrint = true };
+			htmlSettings.RemoveScriptStyleTypeAttribute = false;
+			htmlSettings.RemoveOptionalTags = false;
+			htmlSettings.AttributeQuoteChar = '\'';
+			htmlSettings.RemoveAttributeQuotes = false;
+			htmlSettings.MinifyCss = false;
+			htmlSettings.MinifyCssAttributes = false;
+			htmlSettings.MinifyJsAttributes = false;
+			htmlSettings.MinifyJs = false;
+			return Uglify.Html(text, htmlSettings).Code;
+		}
 		static public string ValueEncode(string sText)
 		{
 			return ValueEncode(sText, false, false);
@@ -4237,16 +4269,16 @@ namespace GeneXus.Utils
 				else
 				{
 #if NETCORE
-                    var basicAuthenticationHeader = GetBasicAuthenticationHeaderValue(cntxt.HttpContext);
-                    if (basicAuthenticationHeader.IsValidBasicAuthenticationHeaderValue)
-                    {
-                        s = basicAuthenticationHeader.UserIdentifier;
-                    }
-                    else
-                    {
+					BasicAuthenticationHeaderValue basicAuthenticationHeader = GetBasicAuthenticationHeaderValue(cntxt.HttpContext);
+					if (basicAuthenticationHeader != null && basicAuthenticationHeader.IsValidBasicAuthenticationHeaderValue)
+					{
+						s = basicAuthenticationHeader.UserIdentifier;
+					}
+					else
+					{
 						if (IsWindowsPlatform)
 							s = System.Security.Principal.WindowsIdentity.GetCurrent().Name;
-                    }
+					}
 #else
 					s = System.Security.Principal.WindowsIdentity.GetCurrent().Name;
 					GXLogging.Debug(log, "UserId= System.Security.Principal.WindowsIdentity.GetCurrent().Name: ", s);
@@ -4278,10 +4310,15 @@ namespace GeneXus.Utils
 #if NETCORE
         private static BasicAuthenticationHeaderValue GetBasicAuthenticationHeaderValue(HttpContext context)
         {
-            var basicAuthenticationHeader = context.Request.Headers["Authorization"]
-                .FirstOrDefault(header => header.StartsWith("Basic", StringComparison.OrdinalIgnoreCase));
-            var decodedHeader = new BasicAuthenticationHeaderValue(basicAuthenticationHeader);
-            return decodedHeader;
+			if (context == null)
+				return null;
+			else
+			{
+				string basicAuthenticationHeader = context.Request.Headers["Authorization"]
+					.FirstOrDefault(header => header.StartsWith("Basic", StringComparison.OrdinalIgnoreCase));
+				var decodedHeader = new BasicAuthenticationHeaderValue(basicAuthenticationHeader);
+				return decodedHeader;
+			}
         }
 #endif
 #if NETCORE
@@ -4353,9 +4390,8 @@ namespace GeneXus.Utils
 				throw ex;
 			}
 		}
-
-		[Obsolete("UserId with string dataSource is deprecated, use UserId((string key, IGxContext cntxt, IDataStoreProvider dataStore) instead", false)]
-		public static string UserId(string key, IGxContext cntxt, string id)
+		[Obsolete("UserId(string key, IGxContext cntxt, string dataSource) with string dataSource is deprecated, use UserId((string key, IGxContext cntxt, IDataStoreProvider dataStore) instead", false)]
+		public static string UserId(string key, IGxContext cntxt, string dataSource)
 		{
 			try
 			{
@@ -4363,7 +4399,7 @@ namespace GeneXus.Utils
 				if (key.ToUpper() == "SERVER" && !(Config.GetValueOf("LOGIN_AS_USERID", out prop) && prop.Equals("1")))
 				{
 					GXLogging.Debug(log, "UserId= user in ConnectionString");
-					IGxDataStore dstore = cntxt.GetDataStore(id);
+					IGxDataStore dstore = cntxt.GetDataStore(dataSource);
 					if (dstore != null)
 						return (dstore.UserId);
 				}
@@ -4521,15 +4557,6 @@ namespace GeneXus.Utils
 			return Entry;
 		}
 
-#if NETCORE
-        private static string WrkSt(IGxContext gxContext, object httpContext)
-
-        {
-        return string.Empty;
-        }
-#else
-
-
 		private static string WrkSt(IGxContext gxContext, HttpContext httpContext)
 		{
 			try
@@ -4544,9 +4571,10 @@ namespace GeneXus.Utils
 				else if (httpContext != null)
 				{
 					if ((Config.GetValueOf("WRKST_COMPATIBILITY", out prop) && prop.Equals("1"))) //WRKST_COMPATIBILITY=YES at config.gx
-						wrkst = httpContext.Server.MachineName;
+						wrkst = httpContext.GetServerMachineName();
 					else
-						wrkst = httpContext.Request.UserHostAddress;
+						wrkst = httpContext.GetUserHostAddress();
+
 					GXLogging.Debug(log, "WrkSt= &HttpRequest.Remoteaddress");
 				}
 				else //Win
@@ -4562,7 +4590,6 @@ namespace GeneXus.Utils
 				throw ex;
 			}
 		}
-#endif
 
 		internal static string NormalizeEncodingName(string enc1)
 		{
@@ -5595,19 +5622,33 @@ namespace GeneXus.Utils
 	{
 		public const string DELIMITER = "/";
 
-		public static string EncodeUrl(string objectName)
+		private static string ReplaceAt(string str, int index, int length, string replace)
 		{
+			return str.Remove(index, Math.Min(length, str.Length - index))
+					.Insert(index, replace);
+		}
+
+		[Obsolete("EncodeUrl is deprecated as it would give unexpected results for some urls. Use EncodeUrlPath instead", false)]
+		public static string EncodeUrl(string objectName) { 
 			if (!Uri.IsWellFormedUriString(objectName, UriKind.RelativeOrAbsolute))
 				return HttpUtility.UrlPathEncode(objectName);
 			return objectName;
 		}
 
+		public static string EncodeUrlPath(string objectNameOrRelativeUrl)
+		{
+			int idx = objectNameOrRelativeUrl.LastIndexOf(StorageUtils.DELIMITER);
+			if (idx > 0)
+			{
+				string objectName = objectNameOrRelativeUrl.Substring(idx + 1);
+				return ReplaceAt(objectNameOrRelativeUrl, idx + 1, objectName.Length, Uri.EscapeDataString(objectName));
+			}
+			return Uri.EscapeDataString(objectNameOrRelativeUrl);
+		}
+
 		public static string DecodeUrl(string url)
 		{
-			string newUrl;
-			while ((newUrl = HttpUtility.UrlDecode(url)) != url)
-				url = newUrl;
-			return url;
+			return HttpUtility.UrlDecode(url);
 		}
 
 		public static String NormalizeDirectoryName(String directoryName)
