@@ -417,8 +417,8 @@ public class GxExternalFileInfo : IGxFileInfo
 	public GxExternalFileInfo(ExternalProvider provider)
     {
         _provider = provider;
-        _name = "";
-        _url = "";
+        _name = string.Empty;
+        _url = string.Empty;
     }
 
     public GxExternalFileInfo(string objectPath, ExternalProvider provider, GxFileType fileType)
@@ -1111,7 +1111,6 @@ public class GxFile
 	}
     public string XsltApply(string xslFileName)
     {
-#if !NETCORE
         XmlReaderSettings readerSettings = new XmlReaderSettings();
         readerSettings.CheckCharacters = false;
         try
@@ -1123,25 +1122,35 @@ public class GxFile
             GXLogging.Warn(log, "XsltApply Error", ex);
             return GxXsltImpl.ApplyOld(FileUtil.NormalizeSource(xslFileName, _baseDirectory), _file.FullName);
         }
-#else
-		return string.Empty;
-#endif
 	}
-#if !NETCORE
 	[MethodImpl(MethodImplOptions.Synchronized)]
     public string HtmlClean()
     {
         try
         {
-            object rawDoc = Assembly.LoadFrom(GXUtil.ProcessorDependantAssembly("NTidy.dll")).CreateInstance("NTidy.TidyDocument");
-            if (File.Exists(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "tidy.cfg")))
-            {
-                rawDoc.GetType().GetMethod("LoadConfig").Invoke(rawDoc, new object[] { Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "tidy.cfg") });
-            }
-            rawDoc.GetType().GetMethod("LoadFile").Invoke(rawDoc, new object[] { _file.FullName });
-            rawDoc.GetType().GetMethod("CleanAndRepair").Invoke(rawDoc, null);
-            String htmlcleaned = (string)rawDoc.GetType().GetMethod("ToString").Invoke(rawDoc, null);
-            return htmlcleaned;
+#if !NETCORE
+			Assembly ntidy=null;
+			try
+			{
+				ntidy =  Assembly.LoadFrom(GXUtil.ProcessorDependantAssembly("NTidy.dll"));
+			}
+			catch (Exception ex)
+			{
+				GXLogging.Warn(log, "Error loading ntidy", ex);
+			}
+			if (ntidy!=null)
+			{
+				object rawDoc = ntidy.CreateInstance("NTidy.TidyDocument");
+				if (File.Exists(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "tidy.cfg")))
+				{
+					rawDoc.GetType().GetMethod("LoadConfig").Invoke(rawDoc, new object[] { Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "tidy.cfg") });
+				}
+				rawDoc.GetType().GetMethod("LoadFile").Invoke(rawDoc, new object[] { _file.FullName });
+				rawDoc.GetType().GetMethod("CleanAndRepair").Invoke(rawDoc, null);
+				return (string)rawDoc.GetType().GetMethod("ToString").Invoke(rawDoc, null);
+			}
+#endif
+			return GXUtil.HTMLClean(ReadAllText(string.Empty));
         }
         catch (Exception ex)
         {
@@ -1149,7 +1158,7 @@ public class GxFile
             return "";
         }
     }
-#endif
+
 	public bool HasExtension()
     {
         return System.IO.Path.HasExtension(this.GetAbsoluteName());
