@@ -80,6 +80,11 @@ namespace GeneXus.Utils
 				return _nullSQLGeography;
 			}
 		}
+		internal static bool IsNull(object instance)
+		{
+			return (bool)ClassLoader.GetPropValue(instance, "IsNull");
+		}
+
 		internal static object STGeometryType(object instance)
 		{
 			return ClassLoader.Invoke(instance, "STGeometryType", null);
@@ -100,7 +105,7 @@ namespace GeneXus.Utils
         internal static bool IsValid(object instance)
         {
             try
-            {
+            {			
                 return ((SqlBoolean)ClassLoader.Invoke(instance, "STIsValid", null)).Value;
             }
             catch (MissingMethodException ex) {
@@ -206,7 +211,6 @@ namespace GeneXus.Utils
                 this.FromJSON(value.ToString());                
             }
         }
-
 
         public Geospatial(object value)
 		{
@@ -326,7 +330,6 @@ namespace GeneXus.Utils
 			}
 		}
 
-
 		public double Longitude
 		{
 			get
@@ -336,7 +339,6 @@ namespace GeneXus.Utils
 				else
 					return SQLGeographyWrapper.Long(_innerValue);
 			}
-
 		}
 
 		public double Latitude
@@ -602,21 +604,31 @@ namespace GeneXus.Utils
             
             
 			try
-			{
+			{			
 				// Sql Server Text
 				_innerValue = SQLGeographyWrapper.Parse(geoText);
-                if ((!SQLGeographyWrapper.IsValid(_innerValue)) && _innerValue !=null )
-                {
-                    _innerValue = SQLGeographyWrapper.MakeValid(_innerValue);
-                }    
-     
-                this.srid = ((SqlInt32)ClassLoader.GetPropValue(_innerValue, "STSrid")).Value;
 
-                this.setGXGeoType(SQLGeographyWrapper.STGeometryType(_innerValue).ToString());
-				if (GeographicType == GeoGraphicTypeValue.Point)
+				if (_innerValue != null && (!SQLGeographyWrapper.IsNull(_innerValue)))
 				{
-					this.Point.Longitude = SQLGeographyWrapper.Long(_innerValue);
-					this.Point.Latitude = SQLGeographyWrapper.Lat(_innerValue);
+					if (!SQLGeographyWrapper.IsValid(_innerValue))
+						_innerValue = SQLGeographyWrapper.MakeValid(_innerValue);
+					
+					if (SQLGeographyWrapper.IsValid(_innerValue))
+					{
+						this.srid = ((SqlInt32)ClassLoader.GetPropValue(_innerValue, "STSrid")).Value;
+						this.setGXGeoType(SQLGeographyWrapper.STGeometryType(_innerValue).ToString());
+						if (GeographicType == GeoGraphicTypeValue.Point)
+						{
+							this.Point.Longitude = SQLGeographyWrapper.Long(_innerValue);
+							this.Point.Latitude = SQLGeographyWrapper.Lat(_innerValue);
+						}
+					}
+					else
+						setNullGeography();
+				}
+				else
+				{					
+						setNullGeography();
 				}
 			}
 			catch (ArgumentException ex)
@@ -631,11 +643,7 @@ namespace GeneXus.Utils
 				}
 				else
 				{
-					// Cannot parse value
-					this.InnerValue = SQLGeographyWrapper.NullSQLGeography;
-					this.geoText = "";
-					this.Point.Longitude = 0;
-					this.Point.Latitude = 0;
+					setNullGeography();
 				}
 			}
 			catch (FormatException ex)
@@ -668,29 +676,17 @@ namespace GeneXus.Utils
 						}
 						catch (Exception)
 						{
-							// Can't convert to geography set as null.
-							_innerValue = SQLGeographyWrapper.NullSQLGeography;
-							this.geoText = "";
-							this.Point.Longitude = 0;
-							this.Point.Latitude = 0;
+							setNullGeography();
 						}
 					}
 					else
 					{
-						// Cannot parse value
-						_innerValue = SQLGeographyWrapper.NullSQLGeography;
-						this.geoText = "";
-						this.Point.Longitude = 0;
-						this.Point.Latitude = 0;
+						setNullGeography();
 					}
 				}
 				else
 				{
-					// Cannot parse value
-					_innerValue = SQLGeographyWrapper.NullSQLGeography;
-					this.geoText = "";
-					this.Point.Longitude = 0;
-					this.Point.Latitude = 0;
+					setNullGeography();
 				}
 			}
 		}
@@ -699,6 +695,15 @@ namespace GeneXus.Utils
         {
 			return this.ToStringSQL("");
         }
+
+		void setNullGeography()
+		{
+			// Cannot parse value
+			_innerValue = SQLGeographyWrapper.NullSQLGeography;
+			this.geoText = "";
+			this.Point.Longitude = 0;
+			this.Point.Latitude = 0;
+		}
 
         public String ToStringESQL()
         {
@@ -885,9 +890,8 @@ namespace GeneXus.Utils
                     object currentRing = ClassLoader.Invoke(geoInstance, "RingN", new object[] { j });
                     SqlInt32 p = (SqlInt32)ClassLoader.Invoke(currentRing, "STNumPoints", null);
 
-                    for (int i = 1; i < p.Value; i++)
+                    for (int i = 1; i <= p.Value; i++)
                     {
-
                         double[] CurrentPoint = new double[2];
                         object currentPoint = ClassLoader.Invoke(currentRing, "STPointN", new object[] { i });
                         CurrentPoint[0] = SQLGeographyWrapper.Long(currentPoint);
