@@ -14,22 +14,42 @@ namespace GeneXus.Application
 	{
 		private Dictionary<string, string> callTargetsByObject = new Dictionary<string, string>();
 		protected IGxContext _Context;
+		bool _isMain;
+		protected bool _isApi;
+
 		public virtual IGxContext context
 		{
 			set { _Context = value; }
 			get { return _Context; }
 			
 		}
+		public bool IsMain
+		{
+			set { _isMain = value; }
+			get { return _isMain; }
+		}
+		public bool IsApiObject
+		{
+			set { _isApi = value; }
+			get { return _isApi; }
+		}
+		public virtual void cleanup() { }
+
 		virtual public bool UploadEnabled() { return false; }
 		public bool IntegratedSecurityEnabled2 { get { return IntegratedSecurityEnabled; } }
 		public GAMSecurityLevel IntegratedSecurityLevel2 { get { return IntegratedSecurityLevel; } }
 		public bool IsSynchronizer2 { get { return IsSynchronizer; } }
 		public string ExecutePermissionPrefix2 { get { return ExecutePermissionPrefix; } }
 
+		public virtual string ServiceExecutePermissionPrefix { get { return string.Empty; } }
+		public virtual string ServiceDeletePermissionPrefix { get { return string.Empty; } }
+		public virtual string ServiceInsertPermissionPrefix { get { return string.Empty; } }
+		public virtual string ServiceUpdatePermissionPrefix { get { return string.Empty; } }
+
 		protected virtual bool IntegratedSecurityEnabled { get { return false; } }
 		protected virtual GAMSecurityLevel IntegratedSecurityLevel { get { return 0; } }
 		protected virtual bool IsSynchronizer { get { return false; } }
-		protected virtual string ExecutePermissionPrefix { get { return ""; } }
+		protected virtual string ExecutePermissionPrefix { get { return String.Empty; } }
 
 		public virtual void CallWebObject(string url)
 		{
