@@ -19,12 +19,13 @@ namespace GeneXus.Application
 		static Regex scheme = new Regex(schemeRegEx, RegexOptions.IgnoreCase);
 		internal static string GetURLRoute(string key, object[] objectParms, string[] parmsName, string scriptPath)
 		{
-			string[] parms = objectParms.Select(p => StringizeParm(p)).ToArray() ;
-			if (PathUtil.IsAbsoluteUrl(key) || key.StartsWith("/") || scheme.IsMatch(key))
+			string[] parms = objectParms.Select(p => StringizeParm(p)).ToArray();
+			if (PathUtil.IsAbsoluteUrl(key) || key.StartsWith("/") || string.IsNullOrEmpty(key) || scheme.IsMatch(key))
 			{
 				if (parms.Length > 0)
 				{
-					return $"{key}{ConvertParmsToQueryString(parms, parmsName)}";
+					string[] noParmsName = Array.Empty<string>();
+					return $"{key}{ConvertParmsToQueryString(parms, noParmsName)}";
 				}
 				else
 				{
