@@ -21,10 +21,7 @@ namespace GeneXus.Procedure
 	public abstract class GXProcedure: GXBaseObject
 	{
 		static readonly ILog log = log4net.LogManager.GetLogger(typeof(GeneXus.Procedure.GXProcedure));
-		protected bool _isMain;
-		protected bool _isApi;
 		public abstract void initialize();
-		public abstract void cleanup();
 
 		protected int handle;
 
@@ -123,47 +120,53 @@ namespace GeneXus.Procedure
             return Encrypt64(GXUtil.GetHash(GeneXus.Web.Security.WebSecurityHelper.StripInvalidChars(value), Cryptography.Constants.SecurityHashAlgorithm), key);
         }
 
-        protected String Encrypt64(String value, String key)
-        {
-            String sRet = "";
-            try
-            {
-                sRet = Crypto.Encrypt64(value, key);
-            }
-            catch (InvalidKeyException)
-            {
-                GXLogging.Error(log, "440 Invalid encryption key");
-            }
-            return sRet;
-        }
+		protected string Encrypt64(string value, string key)
+		{
+			return Encrypt64(value, key, false);
+		}
+		private string Encrypt64(string value, string key, bool safeEncoding)
+		{
+			string sRet = string.Empty;
+			try
+			{
+				sRet = Crypto.Encrypt64(value, key, safeEncoding);
+			}
+			catch (InvalidKeyException)
+			{
+				GXLogging.Error(log, "440 Invalid encryption key");
+			}
+			return sRet;
+		}
+		protected string UriEncrypt64(string value, string key)
+		{
+			return Encrypt64(value, key, true);
+		}
 
-        protected String Decrypt64(String value, String key)
-        {
-            String sRet = "";
-            try
-            {
-                sRet = Crypto.Decrypt64(value, key);
-            }
-            catch (InvalidKeyException)
-            {
-                GXLogging.Error(log, "440 Invalid encryption key");
-            }
-            return sRet;
-        }
+		protected string Decrypt64(string value, string key)
+		{
+			return Decrypt64(value, key, false);
+		}
+		private string Decrypt64(string value, string key, bool safeEncoding)
+		{
+			String sRet = string.Empty;
+			try
+			{
+				sRet = Crypto.Decrypt64(value, key, safeEncoding);
+			}
+			catch (InvalidKeyException)
+			{
+				GXLogging.Error(log, "440 Invalid encryption key");
+			}
+			return sRet;
+		}
+		protected string UriDecrypt64(string value, string key)
+		{
+			return Decrypt64(value, key, true);
+		}
 		public msglist GX_msglist 
 		{
 			get	{ return context.GX_msglist ; }
 			set	{ context.GX_msglist = value;}
-		}
-		public bool IsMain
-		{
-			set	{ _isMain = value; }
-			get	{ return _isMain;  }
-		}
-		public bool IsApiObject
-		{
-			set { _isApi = value; }
-			get { return _isApi; }
 		}
 		public void setContextReportHandler()
 		{
