@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -12,7 +12,6 @@ namespace GeneXus.Notifications
         private static string ID = "GX_PROGRESS_BAR";
         private GXWebNotification notification;
         private GXWebProgressIndicatorInfo info;
-        private bool running;
 
         public GXWebProgressIndicator(IGxContext gxContext)
         {
@@ -22,26 +21,21 @@ namespace GeneXus.Notifications
 
         public void Show()
         {
-            running = true;
             info.Action = "0";
             UpdateProgress();
         }
 
         private void UpdateProgress()
         {
-            if (running)
-            {
-                GXWebNotificationInfo notif = new GXWebNotificationInfo();
-                notif.Id = GXWebProgressIndicator.ID;
-                notif.GroupName = GXWebProgressIndicator.ID;
-                notif.Message = info;
-                notification.Notify(notif);
-            }
+            GXWebNotificationInfo notif = new GXWebNotificationInfo();
+            notif.Id = GXWebProgressIndicator.ID;
+            notif.GroupName = GXWebProgressIndicator.ID;
+            notif.Message = info;
+            notification.Notify(notif);
         }
 
         public void ShowWithTitle(string title)
         {
-            running = true;
             Title = title;
             info.Action = "1";
             UpdateProgress();
@@ -49,7 +43,6 @@ namespace GeneXus.Notifications
 
         public void ShowWithTitleAndDescription(string title, string desc)
         {
-            running = true;
             Title = title;
             Description = desc;
             info.Action = "2";
@@ -60,7 +53,6 @@ namespace GeneXus.Notifications
         {
             info.Action = "3";
             UpdateProgress();
-            running = false;
         }
 
         public string Class
