@@ -4,7 +4,6 @@ using System.IO;
 using System.Collections;
 using System.Threading;
 using System.Text;
-using iTextSharp.text.io;
 using Microsoft.Win32;
 using System.Runtime.InteropServices;
 using System.Globalization;
@@ -21,6 +20,8 @@ using System.Collections.Generic;
 using System.Security;
 using GeneXus;
 using GeneXus.Utils;
+using System.Reflection;
+using GeneXus.Metadata;
 
 namespace com.genexus.reports
 {
@@ -44,7 +45,9 @@ namespace com.genexus.reports
 		private bool fontUnderline;
 		private bool fontStrikethru;
 		private int fontSize;
-		private BaseColor backColor, foreColor;
+
+		//Color for, BaseColor for => Itext5
+		private object backColor, foreColor, templateColorFill;
 		private Stream outputStream = null; 
 											
 		private ParseINI props = new ParseINI();
@@ -73,7 +76,6 @@ namespace com.genexus.reports
 		private BaseFont templateFont;
 		private int templateFontSize;
 		private bool backFill = true;
-		private BaseColor templateColorFill;
 		private int templateAlignment;
 		private int pages = 0;
 		private bool templateCreated = false;
@@ -96,6 +98,7 @@ namespace com.genexus.reports
 		int STYLE_NONE_CONST = 1;
 		int runDirection;
 		int justifiedType;
+		static Assembly iTextAssembly = typeof(Document).Assembly;
 
 		public bool GxOpenDoc(string fileName)
 		{
@@ -667,7 +670,7 @@ namespace com.genexus.reports
 
 				if (backMode != 0)
 				{
-					cb.SetColorFill(new BaseColor(backRed, backGreen, backBlue));
+					ClassLoader.Invoke(cb, "SetColorFill", new object[] { GetColor(backRed, backGreen, backBlue) });
 					cb.FillStroke();
 				}
 				cb.ClosePathStroke();
@@ -706,7 +709,7 @@ namespace com.genexus.reports
 					roundRectangle(cb, x1, y1, w, h,
 						cRadioTL, cRadioTR,
 						cRadioBL, cRadioBR);
-					cb.SetColorFill(new BaseColor(backRed, backGreen, backBlue));
+					ClassLoader.Invoke(cb, "SetColorFill", new object[] { GetColor(backRed, backGreen, backBlue) });
 					cb.FillStroke();
 					cb.SetLineWidth(penAux);
 
@@ -893,8 +896,8 @@ namespace com.genexus.reports
 			this.fontUnderline = fontUnderline;
 			this.fontStrikethru = fontStrikethru;
 			this.fontSize = fontSize;
-			foreColor = new BaseColor(foreRed, foreGreen, foreBlue);
-			backColor = new BaseColor(backRed, backGreen, backBlue);
+			foreColor = GetColor(foreRed, foreGreen, foreBlue);
+			backColor = GetColor(backRed, backGreen, backBlue);
 
 			backFill = (backMode != 0);
 			try
@@ -1067,7 +1070,12 @@ namespace com.genexus.reports
 			{
 				if (!asianFontsDllLoaded)
 				{
-					StreamUtil.AddToResourceSearch(System.Reflection.Assembly.Load("iTextAsian"));
+					Assembly itextAsian = Assembly.Load("iTextAsian");
+					if (IsItext4())
+						ClassLoader.InvokeStatic(iTextAssembly, "iTextSharp.text.pdf.BaseFont", "AddToResourceSearch", new object[] { itextAsian });
+					else
+						ClassLoader.InvokeStatic(iTextAssembly, "iTextSharp.text.io.StreamUtil", "AddToResourceSearch", new object[] { itextAsian });
+
 					asianFontsDllLoaded = true;
 				}
 			}
@@ -1075,6 +1083,10 @@ namespace com.genexus.reports
 			{
 				GXLogging.Debug(log,"LoadAsianFontsDll error", ae);
 			}
+		}
+		private bool IsItext4()
+		{
+			return iTextAssembly.GetName().Version.Major==4;
 		}
 		public void setAsianFont(String fontName, String style)
 		{
@@ -1140,13 +1152,13 @@ namespace com.genexus.reports
 
 			if (printRectangle && (border == 1 || backFill))
 			{
-				GxDrawRect(left, top, right, bottom, border, foreColor.R, foreColor.G, foreColor.B, backFill ? 1 : 0, backColor.R, backColor.G, backColor.B, 0, 0);
+				GxDrawRect(left, top, right, bottom, border, GetColorR(foreColor), GetColorG(foreColor), GetColorB(foreColor), backFill ? 1 : 0, GetColorR(backColor), GetColorG(backColor), GetColorB(backColor), 0, 0);
 			}
 			Font font = new Font(baseFont, fontSize);
 			int arabicOptions = 0;
 			PdfContentByte cb = writer.DirectContent;
 			cb.SetFontAndSize(baseFont, fontSize);
-			cb.SetColorFill(foreColor);
+			ClassLoader.Invoke(cb, "SetColorFill", new object[] { foreColor });
 			sTxt = sTxt.TrimEnd(TRIM_CHARS);
 			float captionHeight = baseFont.GetFontDescriptor(BaseFont.CAPHEIGHT, fontSize);
 			float rectangleWidth = baseFont.GetWidthPoint(sTxt, fontSize);
@@ -1189,8 +1201,7 @@ namespace com.genexus.reports
 			{
 				
 				StyleSheet styles = new StyleSheet();
-				TrueTypeFont ttFont = baseFont as TrueTypeFont;
-				if (ttFont != null)
+				if (IsTrueType(baseFont))
 				{
 					Hashtable locations = GetFontLocations();
 					foreach (string fontName in locations.Keys)
@@ -1220,7 +1231,9 @@ namespace com.genexus.reports
 
 
 				ColumnText Col = new ColumnText(cb);
-				Col.Alignment = ColumnAlignment(alignment);
+				int colAlignment = ColumnAlignment(alignment);
+				if (colAlignment != 0)
+					Col.Alignment = colAlignment;
 				ColumnText simulationCol = new ColumnText(null);
 				float drawingPageHeight = (float)this.pageSize.Top - topMargin - bottomMargin;
 
@@ -1229,17 +1242,23 @@ namespace com.genexus.reports
 				 rightAux + leftMargin,
 				drawingPageHeight - topAux);
 
-				Col.SetSimpleColumn(rect);
-				simulationCol.SetSimpleColumn(rect);
+				SetSimpleColumn(Col, rect);
+				SetSimpleColumn(simulationCol, rect);
 				
 				try
 				{
-					List<IElement> objects = HTMLWorker.ParseToList(new StringReader(sTxt), styles);
-					for (int k = 0; k < objects.Count; ++k)
+					PdfPCell cell = GetPdfCell();
+					PdfPTable t = GetPdfTable(rightAux - leftAux);
+
+					bool firstPar = true;
+					ICollection objects = (ICollection)ClassLoader.InvokeStatic(iTextAssembly, typeof(HTMLWorker).FullName, "ParseToList", new object[] { new StringReader(sTxt), styles });
+					bool drawWithTable = IsItext4() && DrawWithTable(objects);
+					
+					foreach (object element in objects)
 					{
 						if (PageHeightExceeded(bottomAux, drawingPageHeight))
 						{
-							simulationCol.AddElement((IElement)objects[k]);
+							simulationCol.AddElement((IElement)element);
 							simulationCol.Go(true);
 
 							if (simulationCol.YLine < bottomMargin)
@@ -1248,24 +1267,56 @@ namespace com.genexus.reports
 
 								bottomAux -= drawingPageHeight;
 
+								if (drawWithTable)
+								{
+									ClassLoader.Invoke(t, "AddCell", new object[] { cell });
+									t.WriteSelectedRows(0, -1, leftAux + leftMargin, drawingPageHeight - topAux, cb);
+								}
+
 								GxEndPage();
 								GxStartPage();
 								simulationCol = new ColumnText(null);
-								simulationCol.SetSimpleColumn(rect);
-								simulationCol.AddElement((IElement)objects[k]);
+								SetSimpleColumn(simulationCol, rect);
+								simulationCol.AddElement((IElement)element);
 
-								Col = new ColumnText(cb);
-								Col.Alignment = ColumnAlignment(alignment);
-								Col.SetSimpleColumn(rect);
+								if (drawWithTable)
+								{
+									cell = GetPdfCell();
+									t = GetPdfTable(rightAux - leftAux);
+								}
+								else
+								{
+									Col = new ColumnText(cb);
+									if (colAlignment != 0)
+										Col.Alignment = colAlignment;
+									SetSimpleColumn(Col, rect);
+								}
+
 							}
 						}
-
-						Paragraph p = objects[k] as Paragraph;
+						Paragraph p = element as Paragraph;
 						if (p != null)
-							p.Alignment = ColumnAlignment(alignment);
-
-						Col.AddElement(objects[k]);
-						Col.Go();
+						{
+							if (colAlignment != 0)
+								p.Alignment = colAlignment;
+							if (firstPar && drawWithTable)
+							{
+								p.SetLeading(0, 1);
+								firstPar = false;
+							}
+						}
+						if (drawWithTable)
+							cell.AddElement((IElement)element);
+						else
+						{
+							Col.AddElement((IElement)element);
+							Col.Go();
+						}
+					}
+					if (drawWithTable)
+					{
+						ClassLoader.Invoke(t, "AddCell", new object[] { cell });
+						t.WriteSelectedRows(0, -1, leftAux + leftMargin, drawingPageHeight - topAux, cb);
 					}
 				}
 				catch (Exception ex1)
@@ -1312,7 +1363,7 @@ namespace com.genexus.reports
 					else
 						barcode.X = Const.OPTIMAL_MINIMU_BAR_WIDTH_LARGE_FONT;
 
-					Image imageCode = barcode.CreateImageWithBarcode(cb, backFill ? backColor : null, foreColor);
+					Image imageCode = (Image)ClassLoader.Invoke(barcode, "CreateImageWithBarcode", new object[] { cb, backFill ? backColor : null, foreColor });
 					imageCode.SetAbsolutePosition(leftAux + leftMargin, rectangle.Bottom);
 					barcode.BarHeight = rectangle.Height;
 					imageCode.ScaleToFit(rectangle.Width, rectangle.Height);
@@ -1342,7 +1393,7 @@ namespace com.genexus.reports
 							rectangle = new iTextSharp.text.Rectangle(leftAux + leftMargin, (float)this.pageSize.Top - bottomAux - topMargin - bottomMargin, leftAux + leftMargin + rectangleWidth, (float)this.pageSize.Top - topAux - topMargin - bottomMargin);
 							break;
 					}
-					rectangle.BackgroundColor = backColor;
+					ClassLoader.SetPropValue(rectangle, "BackgroundColor", backColor);
 					try
 					{
 						document.Add(rectangle);
@@ -1360,7 +1411,7 @@ namespace com.genexus.reports
 				if (fontUnderline)
 				{
 					GXLogging.Debug(log,"underlineSeparation: " + underlineSeparation);
-					GXLogging.Debug(log,"Kerning: " + baseFont.GetKerning('a', 'b'));
+					GXLogging.Debug(log,"Kerning: " + GetKerning(baseFont, 'a', 'b'));
 
 					underline = new iTextSharp.text.Rectangle(0, 0);
 					switch (alignment)
@@ -1385,7 +1436,7 @@ namespace com.genexus.reports
 								this.pageSize.Top - bottomAux - topMargin - bottomMargin + startHeight - underlineHeight);
 							break;
 					}
-					underline.BackgroundColor = foreColor;
+					ClassLoader.SetPropValue(underline, "BackgroundColor", foreColor);
 					try
 					{
 						document.Add(underline);
@@ -1403,7 +1454,7 @@ namespace com.genexus.reports
 					float strikethruSeparation = lineHeight / 2;
 
 					GXLogging.Debug(log,"underlineSeparation: " + underlineSeparation);
-					GXLogging.Debug(log,"Kerning: " + baseFont.GetKerning('a', 'b'));
+					GXLogging.Debug(log,"Kerning: " + GetKerning(baseFont,'a', 'b'));
 					switch (alignment)
 					{
 						case 1: // Center Alignment
@@ -1426,7 +1477,7 @@ namespace com.genexus.reports
 								this.pageSize.Top - bottomAux - topMargin - bottomMargin + startHeight - underlineHeight + strikethruSeparation);
 							break;
 					}
-					underline.BackgroundColor = foreColor;
+					ClassLoader.SetPropValue(underline, "BackgroundColor", foreColor);
 					try
 					{
 						document.Add(underline);
@@ -1508,6 +1559,68 @@ namespace com.genexus.reports
 				}
 			}
 		}
+
+		private PdfPTable GetPdfTable(float totalWidth)
+		{
+			PdfPTable table = new PdfPTable(1);
+			table.TotalWidth = totalWidth;
+			return table;
+		}
+
+		private PdfPCell GetPdfCell()
+		{
+			PdfPCell cell = new PdfPCell();
+			cell.Border = Rectangle.NO_BORDER;
+			cell.Padding = 0;
+			return cell;
+		}
+
+		private bool DrawWithTable(ICollection objects)
+		{
+			bool allImages = true;
+			foreach (object element in objects)
+			{
+				Paragraph p = element as Paragraph;
+				if (p != null)
+				{
+					foreach (Chunk ch in p.Chunks)
+					{
+						if (!ch.HasAttributes() || ch.Attributes["IMAGE"] == null)
+							return false;
+					}
+				}
+				else
+				{
+					allImages = false;
+				}
+			}
+			return allImages;
+		}
+
+		private int GetColorB(object backColor)
+		{
+			return (int)ClassLoader.GetPropValue(backColor, "B");
+		}
+
+		private int GetColorG(object backColor)
+		{
+			return (int)ClassLoader.GetPropValue(backColor, "G");
+		}
+
+		private int GetColorR(object backColor)
+		{
+			return (int)ClassLoader.GetPropValue(backColor, "R");
+		}
+
+		private int GetKerning(BaseFont baseFont, char v1, char v2)
+		{
+			if (IsItext4())
+				return baseFont.GetKerning(v1, v2);
+			else
+				return (int)ClassLoader.Invoke(baseFont, "GetKerning", new object[] {(int)Char.GetNumericValue(v1), (int)Char.GetNumericValue(v2) });
+			
+		}
+
 		bool PageHeightExceeded(float bottomAux, float drawingPageHeight)
 		{
 			return bottomAux > drawingPageHeight;
@@ -1521,7 +1634,7 @@ namespace com.genexus.reports
 			Col.RunDirection = runDirection;
 			Col.Alignment = alignment;
 			Col.SetLeading(leading, 1);
-			Col.SetSimpleColumn(rect);
+			SetSimpleColumn(Col, rect);
 			Col.AddText(p);
 			Col.Go(true);
 			return Col;
@@ -1553,7 +1666,7 @@ namespace com.genexus.reports
 				Col.SetLeading(0, 1);
 			else
 				Col.SetLeading(leading, 1);
-			Col.SetSimpleColumn(rect);
+			SetSimpleColumn(Col, rect);
 			Col.AddText(p);
 			Col.Go();
 		}
@@ -1746,7 +1859,7 @@ namespace com.genexus.reports
 			{
 				template.BeginText();
 				template.SetFontAndSize(templateFont, templateFontSize);
-				template.SetColorFill(templateColorFill);
+				ClassLoader.Invoke(template, "SetColorFill", new object[] { templateColorFill });
 				switch (templateAlignment)
 				{
 					case 1: // Center Alignment
@@ -2127,6 +2240,25 @@ namespace com.genexus.reports
 			M_bot = bot;
 		}
 
+		private object GetColor(int backRed, int backGreen, int backBlue)
+		{
+			Type color;
+			if (IsItext4())
+				color = iTextAssembly.GetType("iTextSharp.text.Color"); 
+			else
+				color = iTextAssembly.GetType("iTextSharp.text.BaseColor"); 
+			return ClassLoader.CreateInstance(iTextAssembly, color.FullName, new object[] { backRed, backGreen, backBlue });
+		}
+
+		private bool IsTrueType(BaseFont baseFont)
+		{
+			return baseFont.FontType == BaseFont.FONT_TYPE_TT;
+		}
+
+		private void SetSimpleColumn(ColumnText col, Rectangle rect)
+		{
+			col.SetSimpleColumn(rect.Left, rect.Bottom, rect.Right, rect.Top);
+		}
 	}
 
 	public class ParseINI
@@ -2159,7 +2291,7 @@ namespace com.genexus.reports
 			this.filename = Path.GetFullPath(filename);
 			try
 			{
-				var inputFileNameOrTemplate = filename;
+				string inputFileNameOrTemplate = filename;
 				if (!String.IsNullOrEmpty(template) && !File.Exists(filename) && File.Exists(template))
 				{
 					inputFileNameOrTemplate = Path.GetFullPath(template);
