@@ -215,7 +215,9 @@ namespace GeneXus.Cache
 			{
 				return null;
 			}
-			return JsonSerializer.Serialize(o);
+			JsonSerializerOptions opts = new JsonSerializerOptions();
+			opts.Converters.Add(new DBNullConverter());
+			return JsonSerializer.Serialize(o, opts);
 		}
 		[SecurityCritical]
 		static T Deserialize<T>(string value)
@@ -229,43 +231,4 @@ namespace GeneXus.Cache
 			return JsonSerializer.Deserialize<T>(value, opts);
 		}
 	}
-	[SecurityCritical]
-	public class ObjectToInferredTypesConverter : JsonConverter<object>
-	{
-		[SecurityCritical]
-		public override bool CanConvert(Type typeToConvert)
-		{
-			return typeof(object) == typeToConvert;
-		}
-		[SecurityCritical]
-		public override object Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-		{
-			switch (reader.TokenType)
-			{
-				case JsonTokenType.True:
-					return true;
-				case JsonTokenType.False:
-					return false;
-				case JsonTokenType.Number:
-					if (reader.TryGetInt64(out long l))
-						return l;
-					else return reader.GetDouble();
-				case JsonTokenType.String:
-					if (reader.TryGetDateTime(out DateTime datetime))
-						return datetime;
-					else return reader.GetString();
-				default:
-					using (JsonDocument document = JsonDocument.ParseValue(ref reader))
-					{
-						return document.RootElement.Clone().ToString();
-					}
-			}
-		}
-		[SecurityCritical]
-		public override void Write(Utf8JsonWriter writer, object value, JsonSerializerOptions options)
-		{
-			throw new NotImplementedException();
-		}
-	}
-
 }
